Show memory usage actually allocated by user PHP code
The memory_get_usage and memory_get_peak_usage functions have a
$real_usage parameter.  Update the MemoryCollector and PDOCollector
collectors to use a value of false instead of true.  For
MemoryCollector, we allow the user to choose which type of memory usage
to gather.

To recap, when $real_usage is set (and previously this was the
hard-coded default), PHP returns the amount of memory allocated from the
operating system by the PHP memory manager.  This is typically done in
large chunks of memory and is not very granular.

When $real_usage is cleared, PHP returns the amount of memory the user’s
application has actually allocated from the PHP memory manager.

Clearing $real_usage has useful application especially for the
PDOCollector.  As an example, a query that returns only a few rows will
often say that zero bytes were used when $real_usage is set.  That’s
because PHP was able to satisfy the memory requests from memory it had
already allocated from the operating system.  To contrast, the same
query will show an expected few kilobytes of memory usage when
$real_usage is set to false.