Support debug bar collection from command-line (#322)
* Support debug bar collection from command-line

It can be useful to use the PHP Debug Bar from the command line if a
storage layer is configured.  In that scenario, a user would run a PHP
CLI application, then open a web page with the debug bar front-end and
use the open handler to locate the command-line invocation.

However, a problem is that the collect() function assumes that standard
HTTP variables are available - this assumption does not apply when
running from the CLI.  Write a special branch of code for the CLI server
API so that reasonable metadata can be collected for the CLI case.

Without this patch, null method, URI, and IP are logged and it actually
becomes impossible to open the CLI invocation in the open handler - let
alone tell them apart.

* Fix gethostname call in DebugBar::collect

It might fail, and we need to be prepared for that.