RequestIdGeneratorInterface now returns more random values
Implementers of RequestIdGeneratorInterface::generate() need to be sure
that each call to generate() will return a unique ID.

This could fail to happen in the existing implementation if microtime()
returns the same value on two successive generate() calls that come
within the same microsecond.  For example, in a tight ID-generating
loop.  It was unlikely, but now it is practically impossible.

Additionally, implementers should make sure that the return value is not
all-numeric to avoid any mixed data types when IDs are used as keys in
PHP arrays.  (For example, numeric IDs could result in bugs when
array_merge is used.)