Improve formatting of the data set labels (#340)
If the URI ends with a trailing '/', then the existing code would just
show an empty string as the label, which isn't very useful if a page is
making lots of varied AJAX calls to URIs that end with trailing slashes.

Fix this by detecting if the URI ends with a trailing '/'.  If it does,
then we also include the segment of the URI just prior to the
trailing '/'.