Work around stuck memcached entries
Sometimes, we've noticed that the append() call may "succeed" but a
subsequent get() call returns null.  As a result, the debug bar
effectively loses the ability to look up past requests via the open
dialog, as this breaks the find() function.  The only workaround seems
to be to reset the contents of the key/value with a set() call.

It also affects the behavior of our clear() function: a call to delete()
seems to always be required to break this bad state.