truncate the filename part in the dataset title if its too long
fixes #386
i had the same issue, very annoying indeed. firefox doesn't crash. i tried to disable the hardware acceleration in chrome, seemed promising, but did't solve that.

i suppose this PR is not the quality for a permanent fix (sorry for that), but i hope someone more skilled can improve the PR.