Stop TracedStatement throwing warnings for objects
When objects are passed in as SQL parameters if they are able to be cast
to string then we should do so.

If we cannot cast them to string we will short circuit the mb_check_encoding
test so that we never try to use any objects as a parameter of that
function.