allow customization of the memory precision (#530)
currently there is no way to see the raw "peak_usage" bytes in the output. I'd love to have that number available somewhere, but I couldn't figure out how to do that.

this change is the next best thing, as it allows us to override the default precision of "0", so we can get a little more granularity into the memory usage.