Reduce memory usage for consecutive requests with stacked data
Currently, the session stores stacked data, which in turn might contain
the session data if a RequestCollector is active. This leads to
exponential growth in memory usage and rapid performance degradation if
multiple requests add stacked data before the next time the debug bar is
rendered.

We can avoid this be temporarily removing the stacked data from the
session while collection data.